moving isGitHubUrl form WebUtils.js to GitUtils.js
